requirements.txt
# Core
PyMuPDF>=1.23.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# OCR & Layout
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Formula Recognition
pix2tex>=0.1.1

# DeTikZify (Figure -> TikZ)
detikzify @ git+https://github.com/potamides/DeTikZify

# LLM
httpx>=0.25.0

# Web Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Validation
scikit-image>=0.22.0

# Config & Utils
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Async
aiofiles>=23.2.0
